backend/requirements.txt
annotated-doc          0.0.3
annotated-types        0.7.0
anyio                  4.11.0
bcrypt                 5.0.0
click                  8.3.0
colorama               0.4.6
ecdsa                  0.19.1
fastapi                0.121.0
h11                    0.16.0
idna                   3.11
joblib                 1.4.2
mysql-connector-python 9.5.0
numpy                  2.3.4
pandas                 2.3.3
pip                    26.0.1
pyasn1                 0.6.2
pydantic               2.12.4
pydantic_core          2.41.5
python-dateutil        2.9.0.post0
python-dotenv          1.2.1
python-jose            3.5.0
pytz                   2025.2
rsa                    4.9.1
scikit-learn           1.5.2
scipy                  1.17.0
six                    1.17.0
sniffio                1.3.1
starlette              0.49.3
threadpoolctl          3.6.0
typing_extensions      4.15.0
typing-inspection      0.4.2
tzdata                 2025.2
uvicorn                0.38.0